+++ onsiteprint-plugin/blocks/event/block-script-parts/setup-list.js
@@ -1,24 +1,26 @@
 /* ------------------------------------------------------------------------
  #  JS Part Name: Setup List
  *  Block functions included in the Parts Script in the Event Block.
- ?  Updated: 2025-04-13 - 04:33 (Y:m:d - H:i)
- ?  Info: Added new Setup List Script.
+ ?  Updated: 2025-04-14 - 03:55 (Y:m:d - H:i)
+ ?  Info: Added new Lines in the Function (opSetupList).
 ---------------------------------------------------------------------------
  #  TABLE OF CONTENTS:
 ---------------------------------------------------------------------------
 
 	1. 	Import Functions from Scripts
 
-    2. 	Function: Setup List
+    2. 	Function: Setup of the Participant List in Event
 
 ---------------------------------------------------------------------------
  #  1. Import Functions from Scripts
 --------------------------------------------------------------------------- */
 import * as opModuleBasic from '../../../assets/js/inc/basic.js'
-import * as opModuleEvent from '../../assets/js/inc/event/event.js'
+import { opAddParticipant } from './parts.js'
+
+import * as opModuleEvent from '../../../assets/js/inc/event/event.js'
 
 /* ------------------------------------------------------------------------
- #  2. Function: Setup List
+ #  2. Function: Setup of the Participant List in Event
  ?  NB: The function is under construction.
 --------------------------------------------------------------------------- */
 export function opSetupList( debug, block ) {
@@ -26,36 +28,66 @@ export function opSetupList( debug, block ) {
     try {
         
         ///// Get Function Name.
-        var functionName = opSetColumnNumber.name
+        var functionName = opSetupList.name
         
         ///// Set the Debug.
         ////* Set the Parameter If is not defined (true or false).
         if ( debug !== true ) debug = false
         if ( debug ) console.group( `${ functionName }()` )
 
-        ///// Throw Error if the Variables is missing.
-        if ( ! templateId || ! fieldset) throw opModuleBasic.opReturnResponse( true, 404, { 
-                message: `Missing the Template ID or the Fieldset Element!`,
-                line: opModuleBasic.errorLine(),
-                function: functionName
-            } )
-        else {
+        ///// Get Event ID.
+        const eventId = block.getAttribute( 'data-event-id' )
 
-            ///// Get Event ID.
-            const eventId = block.getAttribute( 'data-event-id' )
+        ///// Throw Error if the Event ID is missing.
+        if ( ! eventId ) throw opModuleBasic.opReturnResponse( true, 404, { 
+            message: `Missing the Event ID!`, 
+            line: opModuleBasic.errorLine(),
+            function: functionName
+        } )
 
-            ///// Validate the Response from the Approval.
-            if ( eventId !== false ) throw opModuleBasic.opReturnResponse( false, 404, { 
-                message: `Missing the Event ID!`, 
-                line: opModuleBasic.errorLine(),
-                function: functionName
-            } )
+        ///// Get the Event.
+        //// #NG: Function (opGetEventList) needs to be moved and created as new. 
+        const eventItem = opGetEventList( eventId )
+
+        ///// Validate the Event List Response.
+        if ( eventItem.error !== false ) throw eventItem
+
+        ///// Get the Participant List.
+        const participantList = eventItem.response.eventParticipants
+
+        ///// Get the Event Template.
+        //// #NG: Function (opGetTemplate) needs to be moved and created as new. 
+        const templateResponse = opGetTemplate( eventItem.response.eventTemplate )
+
+        ///// Validate the Event Template Response.
+        if ( templateResponse.error !== false ) throw templateResponse
+
+        ///// Get the Amount of Columns.
+        let columnAmount = templateResponse.response.templateLayoutColumns.charAt(0)
+        
+        ///// Set the Amount of Columns to the Block.
+        block.setAttribute( 'data-column-count', columnAmount )
+
+        ///// Get the elements.
+        let participantListElement = block.querySelector( '.op-participant-rows' )
+
+        ///// For each Participant create Participant Element.
+        for( let i = 0; i < participantList.length; ++i ) {
+
+            opAddParticipant( debug, block, participantListElement, participantList[i], columnAmount )
+
+            /* opAddEventParticipant( debug, block, participantList[i], columnAmount ).then( response => {
+                participantListElement.insertAdjacentHTML( 'afterbegin', response.element )
+            })*/
 
         }
 
+
+
+
         ///// Return the Response.
         return opModuleBasic.opReturnResponse( false, 200, { 
-            message: `The Column Number was set to ${ columnAmount }!`, 
+            message: `The Setup of the Participant List was Created!`, 
             line: opModuleBasic.errorLine(),
             function: functionName
         }, debug )
@@ -63,7 +95,7 @@ export function opSetupList( debug, block ) {
     } catch( errorResponse ) {
 
         ///// Create Error Details.
-        let errorDetails = ( errorResponse.error == true ) ? errorResponse : opModuleBasic.opReturnResponse( false, 400, { 
+        let errorDetails = ( errorResponse.error == true ) ? errorResponse : opModuleBasic.opReturnResponse( true, 400, { 
             message: errorResponse.message,
             line: opModuleBasic.errorLine(),
             function: functionName